Turn collector/statistics inside out
