+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/VirtualFileSystemServices.java
@@ -27,7 +27,7 @@ import org.gradle.api.internal.GradleInternal;
 import org.gradle.api.internal.cache.StringInterner;
 import org.gradle.api.internal.changedetection.state.BuildSessionScopeFileTimeStampInspector;
 import org.gradle.api.internal.changedetection.state.CachingFileHasher;
-import org.gradle.api.internal.changedetection.state.CachingFileHasherStatisticsCollector;
+import org.gradle.api.internal.changedetection.state.CachingFileHasherStatistics;
 import org.gradle.api.internal.changedetection.state.CrossBuildFileHashCache;
 import org.gradle.api.internal.changedetection.state.DefaultResourceSnapshotterCacheService;
 import org.gradle.api.internal.changedetection.state.GradleUserHomeScopeFileTimeStampInspector;
@@ -189,8 +189,8 @@ public class VirtualFileSystemServices extends AbstractPluginServiceRegistry {
     }
 
     private static class GlobalScopeServices {
-        CachingFileHasherStatisticsCollector createCachingFileHasherStatisticsCollector() {
-            return new CachingFileHasherStatisticsCollector();
+        CachingFileHasherStatistics.Collector createCachingFileHasherStatisticsCollector() {
+            return new CachingFileHasherStatistics.Collector();
         }
     }
 
@@ -202,7 +202,7 @@ public class VirtualFileSystemServices extends AbstractPluginServiceRegistry {
         }
 
         FileHasher createCachingFileHasher(
-            CachingFileHasherStatisticsCollector statisticsCollector,
+            CachingFileHasherStatistics.Collector statisticsCollector,
             CrossBuildFileHashCache fileStore,
             FileSystem fileSystem,
             GradleUserHomeScopeFileTimeStampInspector fileTimeStampInspector,
@@ -358,7 +358,7 @@ public class VirtualFileSystemServices extends AbstractPluginServiceRegistry {
             FileSystem fileSystem,
             StreamHasher streamHasher,
             StringInterner stringInterner,
-            CachingFileHasherStatisticsCollector statisticsCollector
+            CachingFileHasherStatistics.Collector statisticsCollector
         ) {
             CachingFileHasher localHasher = new CachingFileHasher(new DefaultFileHasher(streamHasher), cacheAccess, stringInterner, fileTimeStampInspector, "fileHashes", fileSystem, 400000, statisticsCollector);
             return new SplitFileHasher(globalHasher, localHasher, globalCacheLocations);
