+++ subprojects/native/src/main/java/org/gradle/internal/nativeintegration/filesystem/services/FileSystemServices.java
@@ -17,7 +17,7 @@
 package org.gradle.internal.nativeintegration.filesystem.services;
 
 import net.rubygrapefruit.platform.file.PosixFiles;
-import org.gradle.internal.file.FileSystemStatisticsCollector;
+import org.gradle.internal.file.FileSystemStatistics;
 import org.gradle.internal.nativeintegration.filesystem.FileCanonicalizer;
 import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor;
 import org.gradle.internal.nativeintegration.filesystem.FileModeAccessor;
@@ -38,12 +38,12 @@ public class FileSystemServices {
         return newInstanceOrFallback("org.gradle.internal.nativeintegration.filesystem.jdk7.Jdk7FileCanonicalizer", FileSystemServices.class.getClassLoader(), FallbackFileCanonicalizer.class);
     }
 
-    FileSystemStatisticsCollector createFileSystemStatisticsCollector() {
-        return new FileSystemStatisticsCollector();
+    FileSystemStatistics.Collector createFileSystemStatisticsCollector() {
+        return new FileSystemStatistics.Collector();
     }
 
     @SuppressWarnings("UnusedDeclaration")
-    public FileSystem createFileSystem(OperatingSystem operatingSystem, PosixFiles posixFiles, FileMetadataAccessor metadataAccessor, FileSystemStatisticsCollector statisticsCollector) {
+    public FileSystem createFileSystem(OperatingSystem operatingSystem, PosixFiles posixFiles, FileMetadataAccessor metadataAccessor, FileSystemStatistics.Collector statisticsCollector) {
 
         if (operatingSystem.isWindows()) {
             Symlink symlink = newInstanceOrFallback("org.gradle.internal.nativeintegration.filesystem.jdk7.WindowsJdk7Symlink", FileSystemServices.class.getClassLoader(), WindowsSymlink.class);
